Use datagrid component in reports
